CoordinatorServerList now stores only one LogCabin entry id for each server.
We intend to one valid entry in LogCabin log corresponding to each server
that holds all the information about that server. When another coordinator
operation is being done on that server, an entry correspoding to this state
will be appended to the log, and at the end of that operation, the state
entry will be invalidated and a new entry with the new information will be
appended to the log. We need to store the id corresponding to only the
information entry in CoordinatorServerList.