Cancelling active rpcs on infrc didn't release client receive request.
InfRcTransport doesn't send rpc requests if there are no work requests
in the client shared receive queue. The transport increments the
number of in-use receive queue entries when it posts requests for send.

The methods for cancelling rpcs on the transport were missing the
decrement to release the client receive queue entries for transmission
of later rpcs. This had the effect that after enough cancelled active
rpcs, processes silently stop sending rpcs requests and begin queueing
them all.