EnlistServer now logs state and server information to LogCabin.
Before modifying distributed state, it saves the state of the operation.
After the operation has completed, it saves the information about the enlisted
server, invalidates the state entry, and invalidates the information entry
corresponding to the server replaced by the enlisting server.

This will change the way other server-related operations save state as well.