Backups and ReplicaManager use end-to-end footers on replicas.
ReplicaManager queries Segments for the footer that should be
placed on replicas. Backups hold the most recent valid footer for
each replica and use that footer if the replica is closed or used
for recovery.

This patch includes a few hacks which will need to be undone when
the new log module patch is applied.

As a temporary step toward the new log module, Segments no longer
place footers in the segment itself. Footers are fabricated on-demand
for the ReplicaManager. This has led to some breakage of existing
segment/log iterator code which seem to rely on the footer in some
cases.

The checksums are not yet checked before use of replicas during
recovery. This should be implemented once the new log module patch
has been applied.

There is also a bug in this patch where some writes may get
acknowledged to clients before a footer guaranteeing there durability
is acknowledged by the backup. Because fixing this will require
some careful surgery to ReplicaManager it will be pushed as a
separate patch.