ReplicatedSegment::sync() waits for writes to commit on backups.
Fixes issue mentioned in 61e84. In the previous patch some object
might get batched and sent to backups in a single rpc. If that rpc
was larger than batching limit in ReplicaManager then no footer
would be sent along with the rpc. Because 'footer-less' writes
(as of 61e84) aren't considered durable, it is possible an object
which was sent entirely as part of the footer-less rpc could
be sent to and acknowledged by the backup, yet still wouldn't be
found during recovery.

This patch introduces a new state for data enqueued for replication
called 'committed' (in addition to the original 'queued', 'sent',
and 'acked' states). sync() no longer returns until the object
being synced is committed (as opposed to just acked).