Add WARNINGs every 1s in some cases that are prone to silently hang.
Sometimes a master cannot choose a backup because it hasn't been
informed of enough backups by the coordinator. Similarly, an object
may not be synced if one of the backups of the head segment is
not responding and the coordinator hasn't removed it from the
cluster in a timely manner.

These cases now provide occasional output so there is at least some
hope of figuring out where RAMCloud is stuck.

Fixes RAM-421.