Force membership update rpcs to give up.
The new rpc system automatically retries rpcs to servers until those
servers are removed from the server list. This created a deadlock
on the coordinator when two servers crashed at once. One failure
would be detected, and the change to the server list would be broadcast
to the cluster. However, the second crashed server couldn't accept
the update and the coordinator could never give up resending the
update rpc since, in the meantime, the coordinator thread was
blocked and couldn't process the hint server down for the second
server.

Membership updates are send using the asynchronous rpc api now and
are manually timed out if they are taking too long. This is a temporary
fix until the (soon to come) patch which broadcasts cluster
membership updates on a separate thread.