Fix a subtle bug in ServerTracker. We were allowing the enqueueChange method, which is called by the parent ServerList, to resize the internal serverList vector. Unfortunately, doing so invalidates any references into that structure and users of the tracker access it without taking locks.
If you get really super unlucky (and somehow my branch combined
with Stephen Yang's recent changes aligns the planets in a particularly
menacing way), then you might get a race between filling in a new entry
in the old serverList vector's allocation and it copying the contents to
newly resized allocation due to the resize in enqueueChange.

This is what appears to have happened between ServerTracker and
BackupSelector::applyTrackerChanges when additional events were being
simultaneously enqueued.

This is reminiscent of RAM-433, but a lot less obvious.