CoordinatorServerList add() is now split into generateUniqueId() and add().
The older add() generated a unique ServerId and assigned it a new server.
This function was not idempotent since it could generate a new ServerId
each time it was invoked to add the same server.

Now, generateUniqueId() generates a unique ServerId.
The new add() adds a new server with the given ServerId (created by
generateUniqueId()). This function is now idempotent.