ServerDown now logs its own state to LogCabin, and has a recovery path.
serverDown() no longer relies on the calling function to log state before
and after calling it. Any function that wants to force a server out of a cluster
can now acquire a lock for thread safety, and call serverDown().
Currently, it is being used by enlistServer() to remove the server being replaced
and by hintServerDown() after it ensures that the server suspected to be down
is actually down.

If the coordinator crashes during a serverDown() operation, on replay the
new coordinator leader will read StateServerDown from log, and recover via
serverDownRecover() irrespective of the original calling function for serverDown().