Start re-introducing log cleaner metrics to measure performance. The good news is they provide much-needed visibility into the system. The bad news is maintaining the counters tends to clutter the code.
Also, remove the liveness callback. Instead let the relocation
callback move the entry it is wants to. This reduces the number of
callbacks issued by nearly half when most data is alive. It's more
expensive when utilization is low, but in such cases we have more
cycles to spare. It also means not having to pre-append and then
roll back on segments.