Adding function to read only the valid entries from LogCabin log.
Currently, read() function in the LogCabin client API returns all the
entries, including the ones that were invalidated.
Once ongaro implements a cleaner in LogCabin, the read() in LogCabin client
API will return only valid entries and this function will not be needed anymore.
This function is a temporary (and inefficient) work-around to a log cleaner.