Allow waits for segments to be closed on backups to ReplicaManager.
ReplicatedSegment::sync() with no arguments blocks until all data
AND the closed flags have been made durable on backups. This is
useful for the log cleaner which creates segments which must be
manually make completely durable before being patched into the log.

Also Fixes RAM-450 which would have prevented the closing footer
for segments to be sent to backups in some cases.