Fix atomic re-replication for new certificate-based replica scheme.
Since backups no longer have a notion of atomic replication masters
control when replica images commit on backups by withholding
certificates needed to use the replica data until the proper time.

This means when a replica is being recreated the ReplicaManager won't
send a certificate out for that replica until the replica has
completely "caught-up" with the image the master has.