Backups reject open requests if too many (nv) buffers are in use.
Unfortunately, limiting buffers makes it difficult to scale down the
cluster. Because recovery prioritizes replica reads over writes, recovery
needs about enough non-volatile buffers to buffer up all the replicas
generated during recovery. Due to this, the limit is disabled by default,
but it can be easily tuned in the BackupService constructor.

This patch seems to pace recovery to the speed of the storage layer and
should work equally well for pacing normal operation and the log
cleaner. Fixes RAM-467, though it leaves how many buffers is "too many"
to future experimentation.