Recovery from lost open replicas no longer involves log module.
When replicas of open segments may have been lost the replica manager
increments an "epoch" number associated with the segment. It then tries
to contact all existing (unclosed) replicas to tag them with the new
epoch number and recreates a new replica with the new epoch number.
It then updates the replication epoch which consists of the open segment
id and epoch number on the coordinator. During recovery the coordinator
can use the segment id and epoch number to filter out any invalid
replicas.

This decouples the backup recovery logic from the log module because
there is no need to roll over to a new log head with this new
technique. Open segments remain open even after experiencing and
recovering from failures.