Simplify ServerList code since updates can no longer be missed.
The coordinator ensures that all cluster members receive updates in
order. A server that is temporarily unresponsive cannot "skip" over
some server list updates. Because of this a lot of complex code
which would infer events that occurred as part of missed updates
has been dropped.

Additionally, ServerList no longer involves itself in the details of
how events are fires for entries in updates. ServerList blindly
apply updates in order and it is the responsibility of the
centralized coordinator server list to order updates carefully
to preserve any important properties.

Finally, there were a few changes to "tracker" semantics. Trackers can
no longer expect that a server will be ADDED before CRASHED (they may
initially receive a full list with the server already in CRASHED state
and no event are inferred anymore). Also, servers may receive multiple
ADDED events when details about the server have changed. For example,
if the replication id of a server changes it will receive a second
ADDED event. Users of trackers must be prepared for these scenarios
particularly since they affect the allocation/deallocation of
resources placed in the trackers. All existing uses of trackers have
been checked and seem to be safe.

Fixes RAM-453.