Split the startReadingData broadcast rpc to backups into two rpcs. The first litterally only starts the reading of replicas off disk (startReadingData) and the second provides the backup with partitioning information to split up the tablets for each master.
A new addition into the startReadingData rpc is that it now also
has the ability to return tablet profiling data. Right now it's
just a generic string (no actual data), but the framework is there
and is marked with a TODO(syang0).

This completes 1 of 4 steps to the recovery master load balancing
project which are:
  1) Collect Tablet profiling metrics -> Pending for David's TabletMap
  2) Embed Tablet cache of metrics into log -> Pending
  3) Transport Metrics over the RPC system -> Semi-Completed
  4) Redistribute Tablets based on metrics -> Next Step