Nuke the nasty headSegmentIdDuringCleaning nonsense that was put in place to logically order cleaner-generated segments when deciding which objects in the log can possibly be still alive. (This was all needed to avoid recovering dead objects from a previous instance of a tablet after it is migrated away, back again, and a failure occurs).
The new solution is to simply introduce a Log::rollHeadOver() method
that closes the current head and creates a new one with the highest
segment id in the log at that point. The returned position is higher
than anything else in the log at that moment and can be used as the
Log::Position creation time in the coordinator's tablet map.

While here, fix a bug in tablet migration where I was recording the
wrong position for migrated tablets. Previously it would choose the
position immediately _after_ migration completed, meaning that all
migrated data would be considered not part of the tablet during
recovery. Oops.

A unit test was added to address this, but we'll need some seriously
pernicious system tests to really exercise consistency under cleaning,
migration, and recovery.