Make log::sync() thread-safe and have it batch appends to backups when other threads have also appended to the log. This appears to improve small object throughput by about 35-40% with 3 threads in a hacked up parallel MasterService (not committed). I think we're running out of cores to push it much higher.
Also, add another Log::append() method that takes an array of
entries and guarantees that they're all appended to the same
segment atomically. This is used to ensure that tombstones and the
new object are atomically committed to the replicated log. With
a single-threaded master we could just do separate append()s and
then one sync(), but this would lead to races in a multi-threaded
master.

Finally, nuke some dead/superfluous/unused code in Segment.