Clear out followingSegment on ReplicatedSegment::close() if synced.
In testing when numReplicas = 0 the followingSegment couldn't get
cleared since performWrite wouldn't get called. This patch clears
that pointer explicitly in close() if the segment is already
considered durable at the time of the close (for example, when
numReplicas = 0). This also improves an assertion to ensure that
only closed segments are freed rather than just hanging when
that rule is violated.