Shake up of recovery metrics.
First, some attempt is now made for various sections to sum to 100%.
This makes it easier to tell how work shifts between two runs since
for the same section the line items will sum to 100% in both reports.
This has turned out to be something of a pain since many counters
crosscut multiple line items, but some attempt has been made to
account for that.

Second, a grab bag of various metrics some of which may be useful
in the future have been added. There are enough that it may just
be clutter, so we may want to remove some them in the future.