When recovering small objects, recoverSegment() was blowing tons of time updating Metrics counters for every object. The problem is that each counter is an atomic integer, so accessing them is surprisingly expensive.
Instead, we now keep local counters on the stack and update the global counters
once per recovery segment. This saves about 140ms in the 128-byte object case in
RecoverSegmentBenchmark (~10% of the total time)!

This does not address any regression from the sosp submission, as we had the same
problem then. It was relatively worse at that point, however, since recoverSegment
used to run quite a lot faster than it does now (we're still taking about 1.5x as
long).