Cleaning up functions in CoordinatorServerList used to get entries.
The function getReferenceFromServerId() now needs the caller to hold
a lock, and checks that statically to ensure safety.
A function getReferenceFromIndex() has been created to work with an
index number rather than a serverId.

These functions now make the at() functions redundant, hence they have
been removed.

Now, getReferenceFromServerId() and getReferenceFromIndex() is to be
used by functions in  CoordinatorServerList that have already acquired a lock.
The operator[]() functions are to be used by functions in other classes
such that a lock is explicitly acquired here before calling a
getReference method.