Re-enable buffer pooling on backups to boost recovery performance.
There is a bizarre issue with this patch. Before populating the
buffer pool on backups a single buffer is allocated and then freed.
For some reason without this all memory operations during recovery
seem to slow down substantially. Overall, however, this patch
brings us close to the recovery times from SOSP IF one disables
the ReplicatedSegment safety constraints. That issue will be
solved in a forthcoming patch.