Introduce the new SideLog class.
SideLogs look almost exactly like Logs. The only interface difference is
that they have a 'commit()' method, rather than a 'sync()' method. The
key conceptual difference between the two is in append semantics. SideLogs
are designed for appending many entries in bulk and commiting them all at
once with maximum performance. We achieve this in the same way that the
log cleaner operates: segments generated by the SideLog are replicated
without any ordering constraints and are not part of the log until all
are atomically committed.

This new class generalises this cleaner feature. Although it is currently
only used for recovery and tablet migration, it will soon replace the ad-hoc
cleaner mechanisms and may prove useful in future pieces of the system.

SideLog use is very simple, for example:

    SideLog sideLog(&log);
    sideLog.append(..);
    ...
    sideLog.append(...);
    // Appended entries are not yet durable.
    sideLog.commit();
    // Now they are.

Also, while refactoring here, clean up SegmentManager a little bit.