Adding TableManager class that is used by Coordinator and recovery.
CoordinatorService, MasterRecoveryManager, and Recovery all now use
TableManager for all operations related to tablets and tables.
TableManager is built on the Coordinator and not on masters or clients.

It gets created when CoordinatorMain (or MockCluster) are creating the
coordinator. Context holds a pointer to TableManager, which is used
by CoordinatorService when it needs access to TableManager.