Never return from sync on a segment until prior segment is closed.
This fixes issues that arise when the two log digests in each of the
segments are not semantically equivalent. Now calling sync on a
segment up through its header guarantees that when the call returns
the digest in that segment or a later segment will be used during
future recoveries.

This was important for side logs. They inject new segments into the
log digest which contain new data for the log. Now, once the
side log commits the log module only has to roll over the log head
and sync the new log head to commit it safely.

Fixes RAM-489.