Ping RPCs issued by SessionAlarms were misconstructed and always failed.
This would cause long running RPCs to eventually time out even if the
server was responsive to pings. Specifically, the problem is servers now
return a failed not-in-cluster status for pings from servers which are
not in their server list if the supplied server id from the caller is
valid. The extra ping RPC wrapper in SessionAlarm (which exists to avoid
the need for supplying a server id to address it) mistakenly
left the caller id field 0 which is interpreted on the receiving server
as a valid server id which is never in the server list.

Setting the server id to the invalid server id in the ping RPC wrapper
in SessionAlarm fixes the issue.