Never fail fillWithTestData with STATUS_RETRY.
storeObject() returns STATUS_RETRY if the log on the master is out of
space, hoping that the log cleaner will reclaim space.
fillWithTestData just forwards this status on to the client which
causes it to try again.

This adds an ERROR log message when this happens, but doesn't entirely
solve the problem. Killing the master doesn't work because the
fillWithTestData call is then just retried on another server when the
recovery of the first server completes. Since the recovery masters are
configured with even less RAM servers just die in the cluster one
after another.

We probably need a return value from fillWithTestData to tell the client
to back off and notify the user that that the run of recovery.py is a
failure.