Abort recoveries on masters that run out of log space.
recoverSegment() was silently continuing with no warning if an object
couldn't be appended to the log for lack of space. This had the
effect of placing entries in the hash table and completing recovery
even though subsequent accesses to the objects by clients resulted
in 'not found.'

If a recovery master runs out of log space during recovery it now
aborts its share of the recovery. This may be overly aggressive. We may
want to tweak the logic so that it backs off long enough to give the
cleaner time to find more space and then retries. At the very least,
though, this should result in correct behavior.

Changes to create table exacerbated the problem by causing a log head
rollover for each table placed on a master. The recovery.py script now
allocates log space proportional to the number of dummy tables that are
assigned the master which should prevent recovery masters from running
out of log space during test recovery runs.