Generalize the two-phase service initialization scheme that was special-case hacked in for MasterService and BackupService. Also, do enlistment in a separate thread so that RPCs may come in and be processed by services that do not need the ServerId before enlistment completes.
Previously, we would construct Services, enlist, call init() methods
on a few of them, and then enter the dispatch loop. The problem was
that the init() methods had an implicit deadline. If they took longer
than the RPC timeout (say, 500ms), then the dispatch thread wouldn't
get a chance to reply to ping requests and the server would be assumed
to be down.

Now, once enlisted ServiceManager::setServerId is called, which
propagates the id to all services and invokes their internal
initOnceEnlisted() methods. For PingService and MembershipService,
which do not care about ServerIds, this does nothing. They are free to
service RPCs prior to enlistment. MasterService and (for the most part)
BackupService will issue STATUS_RETRY to all RPCs until that method
is called.

Currently, MasterService and BackupService independently track whether
or not their initOnceEnlisted has been fired. This could probably be
pulled into the Service superclass.