Masters abort recovery on coordinator request.
recoveryMasterFinished calls return a bool. The coordinator can tell an
otherwise completed recovery on a recovery master to abort. This is
helpful, for example, when the coordinator rolls over between leader
instances. The new leader doesn't know of the old recoveries. Ongoing
recoveries that suffer a coordinator roll-over should abort on
completion now. Since the tablet map is persisted the new leader will
detect the RECOVERING entries and restart recovery elsewhere.