== Current == Rewrite of the Coordinator's Server List Update mechanism. It is now more understandable with copious amounts of documentation and adds more FT in DCFT for the Coordinator. It is also written in a way that can be easily separated from the main Coordinator (TODO).
== Issues ==
Fixes RAM-484 - The mechanism for updateSlots has been completely rewritten.
Fixes RAM-487 - Assert is removed and isClusterUpToDate has been rewritten.
                The old check was intended to ensure that the updater thread
                never sleeps when there are updates. The new equivalent
                prevents sleeping when there's more work and lockups are
                only transient.

== Future ==
The highest priority is to allow the Coordinator to batch updates. This will
ameliorate some jira issues. The new updater has the framework to allow this
but it has not been completely implemented yet.

The lowest priority is to split the updater into its own class. Again, the
framework has been setup, someone just has to grind through it.