In coordinator implemented expected logging behavior after operation completion.
For Coordinator Operations, log an operation (in LogCabin) as completed after
the update corresponding to this operation has been sent to and acknowledged by
the rest of the cluster, rather than after queuing the updates. Basically,
now the "completed" state in LogCabin corresponds more closely to the actual
completion of the operation.

This has mainly changed the enlistServer() and serverDown() operations, the
corresponding recovery operations, remove() (that is now recoveryCompleted())
and the functionality in pruneUpdates().

Also fixing some latent bugs / improving others things related to / exposed by
these changes.