Trying to make terms related to a crashed servers unambiguous. Basically, a server "crashes", and after its recovery, cluster-wide updates, and cleanup, it is "removed".
More details:

If a server is not responding, it is detected as "crashed" (as opposed to "down"
earlier). This calls hintServerCrashed() in the Coordinator Client (as opposed
to hintServerDown()), which also has correspondingly renamed WireFormat opcode.

Similarly, on the Coordinator (in CSL), the serverCrashed() (as opposed to
serverDown()) starts recovery, sends all required updates to the cluster, and
removes server from server list.

The status of a server in server list is marked as "CRASHED" when it has crashed
and "REMOVE" once it has been successfully recovered (and hence removed from the
cluster), and any data (example, replicas on backups) related to it can also
be removed.