In CoordinatorServerList, handling various states of a crashed server.
When a server crashes, its recovery and cleanup happens as follows:
1. ServerNeedsRecovery: This server's recovery has not been completed, and needs
to be (re-)started when its crash is being handled.
2. ServerCrashed: Its CRASH update is sent out to the cluster and recovery
started if its entry indicates that it needs recovery (set in previous step).
3. ServerRecoveryCompleted: The server's recovery has completed. It now needs
its REMOVE updates to be sent out to the cluster.

Each of these steps are implemented as classes that store data and provide
required functions. They persist needed information to LogCabin, and have
corresponding recovery functions to recover that operation.

This change is motivated by the need to persist server list update version
numbers across coordinator failures. The remainder of the code to handle that
will be committed soon.

(Also fixed some minor latent bugs and style issues discovered along the way.)