Fixed some coordinator bugs (that were resulting in recovery not succeeding).
When coordinator recovers, all the entries corresponding to the servers in the
serverlist have 0 verified version and update version.
This results in the entire server list being sent out to all the servers.
This is not expected behavior (and the servers kill themselves).

It will be too expensive to actually log the most-up-to-date value of verfied
version for each entry during normal operation and then recover that later.
A decent solution (that i've now implemented):
Set these versions for every server entry in server list to be the
ServerListVersion number that was last logged to LogCabin (which was the minimum
verified update number).

Also fixed some coding bugs:
1. CoordinatorServiceRecovery had older entry types -- changed to new ones.
2. CoordinatorServerList modified the LogCabin entryId in local copies of entry
   -- changed to modify in the global CoordinatorServerList entries.