Fixed ParallelRun not to start empty tasks tubs.
Before the code tested to ensure a task tub was populated before calling
wait() on it but not before issuing send(). This was causing a crash
in master recovery on the coordinator when there were more partitions
to recover than recovery masters available. In that case there were
empty tubs representing the partitions which had no recovery master.
Fixes RAM-522.