A new test to test coordinator recovery and basic functioning of RAMCloud.
This test does a series of operations that include starting/crashing
servers, creating/dropping tables and writing/reading objects.
Concurrently, another thread sends RPCs to the coordinator instructing
it to kill itself at specific points in the code. This induces
asynchrony and randomness. The test also checks for consistency based
on the operations already performed and maintains the necessary local
state in various lists.