Supporting multikey objects:
i) This patch replaces the existing Object class to support objects
with variable number of keys - secondary indexes. The format for
an object in the log has also been modified. The format of the object
is almost entirely contained within the Object class. Some APIs/tests
have been modified to reflect the object abstraction.

ii) A subclass of Buffer that understands the new object format is
implemented. It acts as a wrapper to the Object class.

iii) A new variant of the read RPC that reads all the key information
in addition to the object's value. This will be used for indexing
operations.