Removing leakage of indexletId into parts of code that shouldn't know about it.
Each partition of index (indexlet), which is a tree, is (will eventually be)
internally stored as a RAMCloud table. indexletId is this tableId.
However, only the code that does this allocation (or implements related
functionality), should know about it.

Others should refer to an indexlet using the tableId+indexId+key-range.