ObjectManager::removeObject() can now return keys of removed obj.
They keys will be returned in the same format as the keyAndValue
part of an Object, except there will be no value.
This is returned in caller provided buffer. If the caller doesn't
provide a buffer, this is not returned.

This functionality will be used by remove (and multi-remove) in
MasterService to get the keys of removed object and then request
removal of the corresponding index entries.