Write & remove object in ObjectManager now return ref to removed obj.
The writeObject() and removeObject() functions in ObjectManager
now return a pointer to the buffer in log for the object being
overwritten or removed, if any.

This will be used by indexed writes and indexed removes to get data
about (old) index entries to remove.

Note that a similar functionality already existed in removeObject().
This new mechanism is much simpler and (probably) more efficient.