B+-tree implementation using RamCloud objects. This guarantees that the tree is persisted and is recoverable using normal crash recovery. The patch uses the atomic commit for multiple log entries mechanism to guarantee consistency and to avoid garbage on backups.
A subsequent patch will make this the default
B+ tree implementation that the Indexing mechanism
uses.

The code is placed under a new directory under src/
and does not replace the malloc version of B+ tree.
This makes it easier to benchmark both variants
of the B+ tree. As before, by default, this code
is excluded by make check and make tests

NOTE: There are still a few things to address
including possible optimizations and code clean
ups but this can wait until after the deadline
for OSDI '14.