This commit replaces all uses of boost::unordered_multimap with std::unordered_multimap. This involves several changes sated as follows.
1. At this time, we are using gcc-4.4.7 to compile RAMCloud,
which doesn't have 'emplace' as a member function in
std::unordred_multimap (gcc 4.6.0 starts to supports 'emplace').
Therefore, we use 'insert' to implement add item operation, which
is almost the same as 'emplace'. The only difference is that
'insert' performs a move insertion, while 'emplace' advoids that
move operation.

2. In gcc-4.4.7, there is no default hash function for std::pair.
Therefore, if we want to use std::pair<tableID, keyID> (key of
IndexletMap) as a key, we have to implement a user-defined hash
function.