Added a new command line argument to ClusterPerf to specify the number of secondary indexes/object.
Hack in clusterperf.py to make sure that we never wrap around the
host list. This is because if a tablet and its corresponding
indexlet end up on the same server, the server will deadlock. This
also relates to the synchronization or lack thereof of the write
RPC handler in MasterService.