Add checks to Object/IndexServerControl; Add ServerControl RPC.
ObjectServerControl and IndexServerControl now ensure on the server-side that
the responding server owns the object or index in question.  If not, the server
responds with STATUS_UNKNOWN_TABLET/INDEXLET as appropriate.  Both RPC
implementations are unified under a ServerControl RPC.  The ServerControl RPC
may also be used as a building block for any future ServerControlAll RPC.