Improved some indexing related functions in TableManager and Object classes.
1. Improved abstraction for createIndex(), dropIndex() between CoordinatorService
and TableManager classes. Now all of the work (including creating /
dropping backing tables for indexlets) is done by TableManager class and
no information leaks to CoordinatorService.
This has also resulted in monitor-style functions for createTable(),
dropIndex() and dropTable() in TableManager.

2. TableManager::createIndex() now returns true if the index already existed.
This is required for idempotency that is important in case of coordinator crash
and recovery.

3. Changed Object::appendValueToBuffer and Object::appendKeysAndValueToBuffer to
use Buffer* instead of Buffer&.

4. Improved some more documentation and RAMCloud log messages and simplified
some code.