Changed abstraction of lookupIndexKeys and simplifying unit tests.
Changed abstraction of IndexletManager::lookupIndexKeys to accept
request and response headers and rpc as input parameters rather
than individual parameters.
This is done because IndexletManager class had to be indirectly
aware of WireFormat to append the key hashes and next key in the
correct format in the response buffer. This makes the dependency
more explicit (and hence hopefully less bug prone) and simplifies
interface and reduces redundant processing.

This change requires access to MockCluster and RamCloud class
in IndexletManager unit tests. Use this for lookupIndexKeys related
tests. Also simplified the unit tests. It needs some more improvement
to be done in another commit soon.