Improved abstraction for getting indexlet in IndexletManager.
IndexletManager::getIndexlet used to return Indexlet* for indexlet
that had the exact boundaries given.
IndexletManager::lookupIndexlet returns IndexletMap::iterator
for indexlet containing the given index key.

The functions in IndexletManager that needed to find an indexlet
used lookupIndexlet, even if they needed to do additional checks
(similar to getIndexlet) after that. (Note: getIndexlet was not
used by anyone other than unit tests.)

Changed getIndexlet to be an internal monitored function that
returns IndexletMap::iterator. It also now does the checks for
indexlet boundaries and throws InternalError in case of overlapping
boundaries.

The functions in IndexletManager that need an indexlet based on
exact boundaries now use getIndexlet and the ones that need an
indexlet that contains a given key now use lookupIndexlet.