Add zooTest to normal test build, fix minor build issue
Before, zooTest was not part of the normal build at all. Now, it's built
but not executed as part of "make test". (It can't normally be executed
since it depends on a live ZooKeeper cluster.)

zooTest also had a minor build issue caused by the Buffer rewrite, fixed
here (append -> appendCopy).