Add functionality to reassign indexlet and backing table on the coordinator, AND Modify Table protobuf to store normal index-related metadata, AND Fix other lint / naming issues.
Change metadata on coordinator to reflect change of ownership
and contact the old owner to drop ownership and the new server to take
ownership. This is also logged in external storage.

Modify Table protobuf (that is logged in external storage) to have
messages to store normal index metadata (though this is not yet being used
— to be done in a future commit), as well as information about the indexlet
(and corresponding backing table) being reassigned.