Fixes RAM-616; it was a problem with the test itself.
In MembershipServiceTest.updateServerList_multi, there was a line to
disable the ServerListUpdater thread (source.haltUpdater()) which
would cause updates to accumulate in source.updates. That is what
we want. However, right after that line, the test constructs a
CoordinatorService object and the constructor for that calls BACK
into the CoordinatorServerList to restart the updating thread.
And thus, we had this timing issue whereby sometimes the updater
thread would deplete the queue and other times where the test would
get to checking it first. The fix was to move the haltUpdater()
call until after constructing the CoordinatorService.