Add getBackupConfig and getMasterConfig CoordinatorClient RPCs.
First step to centralizing cluster configurations (server runtime
configuration) on the coordinator. Goal is to have default configurations set
on the coordinator and used by all servers in the cluster (instead of setting
the configurations individually for each server). GetBackupConfigRpc and
GetMasterConfigRpc allow servers to request the default configuration from the
coordinator for the backup and master configurations respectively. This commit
adds these RPCs to server startup. However, the modifications to set the
defaults on the coordinator (RAM-671) and use the results on the servers is
left as future work (RAM-672).