What's included: 1) Implementation of RAMCloud ARP Cache for resolving next hop MAC addresses.     (required for layer 3 networking.) 2) Implementation SolarFlareDriver for layer 3 (ie. IP) networking in     RAMCloud. 3) Implementation of zero copy functionality in SolarFlare driver.
This commit contains the ArpCache.{h,cc} files that provide RAMCloud with local
ARP cache, a fast way to resolve next hop MAC address in a layer 3 network.
Also zero copy implementation of SolarFlareDriver is provided in this commit.

Note:
DO NOT use the zero copy version of the driver code for now.
It is an irony that zero copy version of the driver is much slower than the
previous version that copies the received and transmit buffers from the RX and
TX ring to higher layer transport code. The reason is that enabling IO-MMU and
SR-IOV (that are required for zero copy) has significant overhead for
SolarFlare. Currently this version of the zero copy code is here just as a
reference implementation for possible future use.